Reorganize app code
